% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gcfit}
\alias{gcfit}
\title{Creates an object of class gcfit.}
\usage{
gcfit(gc_vals, log_mod, data_t, data_n)
}
\arguments{
\item{gc_vals}{An object of class gcvals that contains the summarized
metrics from fitting the growth model to a set of
experimental observations. This is where the fitness proxy
parameters can be found. See \code{\link{gcvals}} for
more information the information found in this object.}

\item{log_mod}{An object of class nlsModel that contains the results
of fitting the logistic growth model to the data}

\item{data_t}{A numeric vector of times}

\item{data_n}{A numeric vector of cell count or absorbance readings}
}
\value{
An object of class gcfit, which is a list of three objects,
                 that combines the parameters (vals = gc_vals,
                 model = log_mod, data = list(data_t, data_n))
}
\description{
This is a constructor function for the "gcfit" class. This class is most
often obtained as the return value when calling
\code{\link{SummarizeGrowth}}.
}
