% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{growthcurves}
\alias{growthcurves}
\title{Growth curve data from Escherichia coli.}
\format{A data frame with 416 observations and 4 variables:
\describe{
  \item{t}{time, in hours}
  \item{N1}{absorbance readings of well 1}
  \item{N2}{absorbance readings of well 2}
  \item{N3}{absorbance readings of well 3}
}}
\usage{
growthcurves
}
\description{
A dataset containing absorbance measurements over time of E. coli
growing in a plate reader for 1 day. The growth curves for three wells are
included. The background absorbance readings have already been subtracted
from the values by subtracting the minimum absorbance reading in a given well
from the rest of the measurements in that well.
}
\keyword{datasets}

