% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{growthdata}
\alias{growthdata}
\title{Simulated growth curve data}
\format{A data frame with 145 observations and 97 variables:
\describe{
  \item{time}{time, in hours}
  \item{A1}{absorbance readings of well A1}
  \item{A2}{absorbance readings of well A2}
  \item{A3}{absorbance readings of well A3}
  \item{A4}{absorbance readings of well A4}
  \item{A5}{absorbance readings of well A5}
  \item{A6}{absorbance readings of well A6}
  \item{A7}{absorbance readings of well A7}
  \item{A8}{absorbance readings of well A8}
  \item{A9}{absorbance readings of well A9}
  \item{A10}{absorbance readings of well A10}
  \item{A11}{absorbance readings of well A11}
  \item{A12}{absorbance readings of well A12}
  \item{B1}{absorbance readings of well B1}
  \item{B2}{absorbance readings of well B2}
  \item{B3}{absorbance readings of well B3}
  \item{B4}{absorbance readings of well B4}
  \item{B5}{absorbance readings of well B5}
  \item{B6}{absorbance readings of well B6}
  \item{B7}{absorbance readings of well B7}
  \item{B8}{absorbance readings of well B8}
  \item{B9}{absorbance readings of well B9}
  \item{B10}{absorbance readings of well B10}
  \item{B11}{absorbance readings of well B11}
  \item{B12}{absorbance readings of well B12}
  \item{C1}{absorbance readings of well C1}
  \item{C2}{absorbance readings of well C2}
  \item{C3}{absorbance readings of well C3}
  \item{C4}{absorbance readings of well C4}
  \item{C5}{absorbance readings of well C5}
  \item{C6}{absorbance readings of well C6}
  \item{C7}{absorbance readings of well C7}
  \item{C8}{absorbance readings of well C8}
  \item{C9}{absorbance readings of well C9}
  \item{C10}{absorbance readings of well C10}
  \item{C11}{absorbance readings of well C11}
  \item{C12}{absorbance readings of well C12}
  \item{D1}{absorbance readings of well D1}
  \item{D2}{absorbance readings of well D2}
  \item{D3}{absorbance readings of well D3}
  \item{D4}{absorbance readings of well D4}
  \item{D5}{absorbance readings of well D5}
  \item{D6}{absorbance readings of well D6}
  \item{D7}{absorbance readings of well D7}
  \item{D8}{absorbance readings of well D8}
  \item{D9}{absorbance readings of well D9}
  \item{D10}{absorbance readings of well D10}
  \item{D11}{absorbance readings of well D11}
  \item{D12}{absorbance readings of well D12}
  \item{E1}{absorbance readings of well E1}
  \item{E2}{absorbance readings of well E2}
  \item{E3}{absorbance readings of well E3}
  \item{E4}{absorbance readings of well E4}
  \item{E5}{absorbance readings of well E5}
  \item{E6}{absorbance readings of well E6}
  \item{E7}{absorbance readings of well E7}
  \item{E8}{absorbance readings of well E8}
  \item{E9}{absorbance readings of well E9}
  \item{E10}{absorbance readings of well E10}
  \item{E11}{absorbance readings of well E11}
  \item{E12}{absorbance readings of well E12}
  \item{F1}{absorbance readings of well F1}
  \item{F2}{absorbance readings of well F2}
  \item{F3}{absorbance readings of well F3}
  \item{F4}{absorbance readings of well F4}
  \item{F5}{absorbance readings of well F5}
  \item{F6}{absorbance readings of well F6}
  \item{F7}{absorbance readings of well F7}
  \item{F8}{absorbance readings of well F8}
  \item{F9}{absorbance readings of well F9}
  \item{F10}{absorbance readings of well F10}
  \item{F11}{absorbance readings of well F11}
  \item{F12}{absorbance readings of well F12}
  \item{G1}{absorbance readings of well G1}
  \item{G2}{absorbance readings of well G2}
  \item{G3}{absorbance readings of well G3}
  \item{G4}{absorbance readings of well G4}
  \item{G5}{absorbance readings of well G5}
  \item{G6}{absorbance readings of well G6}
  \item{G7}{absorbance readings of well G7}
  \item{G8}{absorbance readings of well G8}
  \item{G9}{absorbance readings of well G9}
  \item{G10}{absorbance readings of well G10}
  \item{G11}{absorbance readings of well G11}
  \item{G12}{absorbance readings of well G12}
  \item{H1}{absorbance readings of well H1}
  \item{H2}{absorbance readings of well H2}
  \item{H3}{absorbance readings of well H3}
  \item{H4}{absorbance readings of well H4}
  \item{H5}{absorbance readings of well H5}
  \item{H6}{absorbance readings of well H6}
  \item{H7}{absorbance readings of well H7}
  \item{H8}{absorbance readings of well H8}
  \item{H9}{absorbance readings of well H9}
  \item{H10}{absorbance readings of well H10}
  \item{H11}{absorbance readings of well H11}
  \item{H12}{absorbance readings of well H12}
}}
\usage{
growthdata
}
\description{
A dataset containing absorbance measurements over time of microbes
growing in a plate reader for 1 day. The growth curves for a whole plate are
included.
}
\keyword{datasets}
