% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NAtT}
\alias{NAtT}
\title{Number of Cells at Time t}
\usage{
NAtT(k, n0, r, t)
}
\arguments{
\item{k}{The carrying capacity}

\item{n0}{The initial population size (absorbance or individuals)}

\item{r}{The exponential "growth rate"}

\item{t}{The time at which you want to know N}
}
\value{
The number of cells, or N, at time t
}
\description{
This function gives the number of cells or absorbance (N) at time t when
the parameters to the logistic equation are K, N0, and r.
}
